Run the reading that names the fork, instead of naming it again in the test
The pair the issue asks about is what the reading names against what the lowering
builds, and the test named one side itself: it called the comprehension for the
fork and compared that with the tree, so the reader whose numbering it was about
never ran. Reversing that reader left it green.

Taking the answer from whoever owns it is the whole of the fix. Both are run now
and their answers put together, over a helper whose guards each rest on a
different supplied rule -- so what the reading says about a fork names which
guard it read, and a numbering off by a place is told from a disagreement about
the set. Breaking either side reddens it.

That makes the older test's first half a second way of asking the same question,
and the halves it alone held belong to the comprehension rather than to the
lowering: a fork is one of the guards written, and the part before the first
would answer with the obligation the construct itself is owed for.
